+++ provision/project.sh
@@ -1,0 +1,6 @@
+# Put any project-specific provisioning steps here
+
+# Available variables:
+#PROJECT_NAME="$1"
+#BUILD_MODE="$2"
+#DEBUG="$3"